backend/pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fastapi-production-template"
version = "1.0.0"
description = "Production-ready FastAPI template with async SQLite, JWT auth, file uploads, background tasks, and Docker."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }

dependencies = [
    # Web framework
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    # ORM + Schemas
    "sqlmodel>=0.0.21",
    # Async SQLite driver
    "aiosqlite>=0.20",
    # Migrations
    "alembic>=1.13",
    # Auth
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7.4",
    # Config
    "pydantic-settings>=2.4",
    "python-dotenv>=1.0",
    # Rate limiting
    "slowapi>=0.1.9",
    # Background tasks
    "arq>=0.26",
    # File uploads
    "python-multipart>=0.0.9",
    # HTTP client (LLM provider calls)
    "httpx>=0.28",
    # PDF extraction
    "pdfplumber>=0.11",
    "pymupdf>=1.24",
    # AI / LLM
    "google-genai>=1.3",
    "email-validator>=2.0.0",
    "bcrypt>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "black>=24",
    "mypy>=1.10",
]

# ── Ruff ──────────────────────────────────────────────────────────────────────
[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
target-version = "py312"
line-length = 100
fix = true

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "SIM",  # flake8-simplify
    "TID",  # flake8-tidy-imports
    "RUF",  # Ruff-specific rules
    "ASYNC",# flake8-async
]
ignore = [
    "E501",  # Line length enforced by Black
    "B008",  # Do not perform function calls in default arguments (needed by Depends)
]

[tool.ruff.lint.isort]
known-first-party = ["app", "tests"]

# ── Black ─────────────────────────────────────────────────────────────────────
[tool.black]
target-version = ["py312"]
line-length = 100

# ── Mypy ──────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

# ── Pytest ─────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore::pytest.PytestUnraisableExceptionWarning",
]

# ── Coverage ──────────────────────────────────────────────────────────────────
[tool.coverage.run]
source = ["app"]
omit = ["app/db/init_db.py", "app/workers/*"]

[tool.coverage.report]
show_missing = true
fail_under = 80
